Extended support for 16 bits addressing
This commit will add the support of 8 and 16 bit address EEPROM read/write fru .

Tested :
all 8 and 16 bit address EEPROM fru getting list Successfully.

dependent commit:-
https://gerrit.openbmc.org/c/openbmc/entity-manager/+/58634

+++ dbus-sdr/storagecommands.cpp
@@ -112,11 +112,11 @@ constexpr static const char* chassisTypeMainServer = "17";
 constexpr static const uint8_t deassertionEvent = 0x80;
 
 static std::vector<uint8_t> fruCache;
-static uint8_t cacheBus = 0xFF;
+static uint16_t cacheBus = 0xFF;
 static uint8_t cacheAddr = 0XFF;
 static uint8_t lastDevId = 0xFF;
 
-static uint8_t writeBus = 0xFF;
+static uint16_t writeBus = 0xFF;
 static uint8_t writeAddr = 0XFF;
 
 std::unique_ptr<phosphor::Timer> writeTimer = nullptr;
